log added to services clases

+++ src/main/java/com/example/gym/facade/GymFacade.java
@@ -30,7 +30,7 @@ public class GymFacade {
         return trainerService.updateTrainerProfile(trainer);
     }
 
-    public Optional<Trainer> getTrainer(int id) {
+    public Optional<Trainer> getTrainer(Long id) {
         return trainerService.selectTrainerProfile(id);
     }
